🎨 Separate manage_receipt windows
