+++ src/presentation/telegram/manage_receipt/__init__.py
@@ -1,0 +1,1 @@
+from .dialog import manage_receipt_dialog as manage_receipt_dialog
