+++ src/presentation/telegram/manage_receipt/common/get_receipt_id.py
@@ -1,0 +1,10 @@
+from typing import TYPE_CHECKING, cast
+
+from src.domain.value_objects import ReceiptID
+
+if TYPE_CHECKING:
+    from aiogram_dialog import DialogManager
+
+
+def get_receipt_id(dialog_manager: DialogManager) -> ReceiptID:
+    return cast(ReceiptID, dialog_manager.start_data["receipt_id"])  # type: ignore[call-overload, index]
